CLN: move align to its own file
